refactor of test files to separate nav testing
